fixes for export/import with not filestore; addition of keyid to area
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@174 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart/src/org/wcs/smart/internal/ca/export/DataStoreDataExporter.java
@@ -54,7 +54,9 @@ public class DataStoreDataExporter implements ICaDataExporter {
 		File filestore = new File(exportEngine.getExportLocation() + File.separator + CaExporter.FILESTORE_DIR);
 		SmartUtils.createDirectory(filestore);
 		File filestoreLocation = new File(exportEngine.getConservationArea().getFileDataStoreLocation());
-		FileUtils.copyDirectory(filestoreLocation, filestore);
+		if (filestoreLocation.exists()){
+			FileUtils.copyDirectory(filestoreLocation, filestore);
+		}
 
 	}
